# src/main/java/ru/adel/if_else_task_2/public_interface/region/dto/RegionResponseDto.java
package ru.adel.if_else_task_2.public_interface.region.dto;

import lombok.Builder;

@Builder
public record RegionResponseDto (
        Long id,
        String name,
        String parentRegion,
        Long regionType,
        Double latitude,
        Double longitude
) { }