# src/main/resources/openapi/weather-app.yaml
openapi: 3.0.2
info:
  title: Weather application
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Development server
security:
  - cookieAuth: []
paths:
...
  /accounts/{accountId}:
    get:
      summary: Получение аккаунта
      tags:
        - Account
      operationId: getAccount
      parameters:
        - name: accountId
          in: path
          description: Идентификатор аккаунта пользователя
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
            example: 1
      responses:
        '200':
          description: Запрос успешно выполнен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Неверные авторизационные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Аккаунт с таким accountId не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Изменение аккаунта
      tags:
        - Account
      operationId: updateAccount
      parameters:
        - name: accountId
          in: path
          description: Идентификатор аккаунта пользователя
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAccountRequest'
      responses:
        '200':
          description: Запрос успешно выполнен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Запрос от неавторизованного аккаунта | Неверные авторизационные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Обновление не своего аккаунта | Аккаунт не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Аккаунт с таким email уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Удаление аккаунта
      tags:
        - Account
      operationId: deleteAccount
      parameters:
        - name: accountId
          in: path
          description: Идентификатор аккаунта пользователя
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
            example: 1
      responses:
        '200':
          description: Запрос успешно выполнен
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Запрос от неавторизованного аккаунта | Неверные авторизационные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Удаление не своего аккаунта | Аккаунт с таким accountId не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /region/{regionId}:
    get:
      summary: Получение информации о регионе
      tags:
        - Region
      operationId: getRegion
      parameters:
        - name: regionId
          in: path
          description: Идентификатор региона
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
            example: 1
      responses:
        '200':
          description: Запрос успешно выполнен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegionResponse'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Неверные авторизационные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Регион с таким regionId не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Изменение региона
      tags:
        - Region
      operationId: updateRegion
      parameters:
        - name: regionId
          in: path
          description: Идентификатор региона
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegionRequest'
      responses:
        '200':
          description: Запрос успешно выполнен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegionResponse'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Неверные авторизационные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Регион с таким regionId не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Регион с такими latitude и longitude уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Удаление региона
      tags:
        - Region
      operationId: deleteRegion
      parameters:
        - name: regionId
          in: path
          description: Идентификатор региона
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
            example: 1
      responses:
        '200':
          description: Запрос успешно выполнен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegionResponse'
        '400':
          description: Ошибка валидации | Регион является родительским для другого региона
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Неверные авторизационные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Регион с таким regionId не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /region:
    post:
      summary: Добавление региона
      tags:
        - Region
      operationId: createRegion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegionRequest'
      responses:
        '201':
          description: Запрос успешно выполнен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegionResponse'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Неверные авторизационные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Регион с такими latitude и longitude уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /region/weather/forecast/{forecastId}:
    get:
      summary: Получение информации о прогнозе погоды
      tags:
        - WeatherForecast
      operationId: getWeatherForecast
      parameters:
        - name: forecastId
          in: path
          description: Идентификатор прогноза
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
            example: 1
      responses:
        '200':
          description: Запрос успешно выполнен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeatherForecastResponse'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Неверные авторизационные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Прогноза с таким forecastld не существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Изменение прогноза погоды
      tags:
        - WeatherForecast
      operationId: updateWeatherForecast
      parameters:
        - name: forecastId
          in: path
          description: Идентификатор прогноза
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWeatherForecastRequest'
      responses:
        '200':
          description: Запрос успешно выполнен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeatherForecastResponse'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Неверные авторизационные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Прогноза с таким forecastld не существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Удаление прогноза погоды
      tags:
        - WeatherForecast
      operationId: deleteWeatherForecast
      parameters:
        - name: forecastId
          in: path
          description: Идентификатор прогноза
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
            example: 1
      responses:
        '200':
          description: Запрос успешно выполнен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeatherForecastResponse'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Неверные авторизационные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Прогноза с таким forecastld не существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /region/weather/forecast:
    post:
      summary: Добавление прогноза погоды
      tags:
        - WeatherForecast
      operationId: createForecast
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WeatherForecastRequest'
      responses:
        '200':
          description: Запрос успешно выполнен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeatherForecastResponse'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Неверные авторизационные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Региона с таким regionId не существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: id
  schemas:
    ErrorResponse:
      type: object
      required:
        - status
        - message
        - timestamp
        - path
      properties:
        status:
          type: integer
          example: 404
          description: Http status code
        message:
          type: string
          description: Error message
        timestamp:
          type: string
          format: date-time
          description: Error occurrence time
        path:
          type: string
          description: URI path
        details:
          type: object
          description: Error details
          additionalProperties:
            type: string
...
    AccountResponse:
      type: object
      required:
        - id
        - firstName
        - lastName
        - email
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
          example: Ivan
          description: Имя пользователя
        lastName:
          type: string
          example: Ivanov
          description: Фамилия пользователя
        email:
          type: string
          format: email
          example: user@mail.ru
          description: Адрес электронной почты
    UpdateAccountRequest:
      type: object
      required:
        - firstName
        - lastName
        - email
        - password
      properties:
        firstName:
          type: string
          example: Ivan
          description: Новое имя пользователя
        lastName:
          type: string
          example: Ivanov
          description: Новая фамилия пользователя
        email:
          type: string
          format: email
          example: user@mail.ru
          description: Новый адрес электронной почты
        password:
          type: string
          format: password
          description: Пароль от аккаунта пользователя
    RegionRequest:
      type: object
      required:
        - name
        - latitude
        - longitude
      properties:
        name:
          type: string
          description: Название региона
        parentRegion:
          type: string
          description: Название родительского региона
        regionType:
          type: integer
          format: int64
          description: Идентификатор типа региона
        latitude:
          type: number
          format: double
          description: Координаты широты
        longitude:
          type: number
          format: double
          description: Координаты долготы
    RegionResponse:
      type: object
      required:
        - id
        - name
        - latitude
        - longitude
      properties:
        id:
          type: integer
          format: int64
          description: Идентификатор региона
        name:
          type: string
          description: Название региона
        parentRegion:
          type: string
          description: Название родительского региона
        regionType:
          type: integer
          format: int64
          description: Идентификатор типа региона
        accountId:
          type: integer
          format: int64
          description: Аккаунт внесший данные о регионе
        latitude:
          type: number
          format: double
          description: Координаты широты
        longitude:
          type: number
          format: double
          description: Координаты долготы
...
    WeatherCondition:
      type: string
      description: Текущее состояние погоды
      enum:
        - CLEAR
        - CLOUDY
        - RAIN
        - SNOW
        - FOG
        - STORM
...
    WeatherForecastResponse:
      type: object
      required:
        - id
        - dateTime
        - temperature
        - weatherCondition
        - regionId
      properties:
        id:
          type: integer
          format: int64
          description: Уникальный идентификатор прогноза погоды
        regionId:
          type: integer
          format: int64
          description: Идентификатор региона, для которого сделан прогноз
        dateTime:
          type: string
          format: date-time
          description: Дата и время прогноза в формате ISO-8601
        temperature:
          type: number
          format: float
          description: Прогнозируемая температура, °C
        weatherCondition:
          $ref: '#/components/schemas/WeatherCondition'
    UpdateWeatherForecastRequest:
      type: object
      required:
        - temperature
        - weatherCondition
        - dateTime
      properties:
        temperature:
          type: number
          format: float
          description: Новая прогнозируемая температура, °C
        weatherCondition:
          $ref: '#/components/schemas/WeatherCondition'
        dateTime:
          type: string
          format: date-time
          description: Новая дата и время прогноза в формате ISO-8601
    WeatherForecastRequest:
      type: object
      required:
        - regionId
        - dateTime
        - temperature
        - weatherCondition
      properties:
        regionId:
          type: integer
          format: int64
          minimum: 1
          description: Идентификатор региона, для которого делается прогноз
        dateTime:
          type: string
          format: date-time
          description: Дата и время, на которые делается прогноз в формате ISO-8601
        temperature:
          type: number
          format: float
          description: Прогнозируемая температура, °C
        weatherCondition:
          $ref: '#/components/schemas/WeatherCondition'